model load

+++ output_text.py
@@ -4,7 +4,7 @@ import time
 
 print("start")
 tokenizer = AutoTokenizer.from_pretrained("EleutherAI/gpt-j-6B", cache_dir="/scratch/network/jmonas/.cache/")
-model = AutoModelForCausalLM.from_pretrained("EleutherAI/gpt-j-6B",torch_dtype=torch.float16, cache_dir="/scratch/network/jmonas/.cache/")
+model = AutoModelForCausalLM.from_pretrained("EleutherAI/gpt-j-6B",cache_dir="/scratch/network/jmonas/.cache/")
 # device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 print("done")
 # print("device selection")
